requirement.txt
# Upgrade pip: pip install -i https://pypi.tuna.tsinghua.edu.cn/simple --upgrade pip
# Install with: pip install -i https://pypi.tuna.tsinghua.edu.cn/simple -r requirements.txt

# Wheel
wheel >= 0.30.0

# Django
django >= 2.2.0
django-cors-headers >= 2.1.0
django-daterange-filter >= 1.3.0
django-extensions >= 1.9.7
django-import-export >= 0.5.1
django-tinymce >= 2.6.0
djangorestframework >= 3.7.3
djangorestframework-jwt >= 1.11.0
django-oauth-toolkit >= 1.2.0

bcrypt >= 3.1.4


# ES
elasticsearch >= 6.2.0

mysqlclient >= 1.3.14

xlrd >= 1.1.0

python-dateutil >= 2.7.2

uwsgi >= 2.0.17

# jieba
jieba >= 0.39

# lxml
lxml >= 4.2.5

# fuzzywuzzy
fuzzywuzzy >= 0.17.0

# requests
requests >= 2.20.0

# beautifulsoup
beautifulsoup4 >= 4.6.3

# baidu AI AipNlp
baidu-aip >= 2.2.13.0

# Pillow
Pillow >= 5.0.0

# python-Levenshtein
python-Levenshtein >= 0.12.0

xpinyin >= 0.5.6
